+++ src/main/java/subway/application/LineService.java
@@ -3,9 +3,9 @@ package subway.application;
 import org.springframework.stereotype.Service;
 import subway.dao.LineEntity;
 import subway.dao.SectionEntity;
-import subway.domain.Line;
-import subway.domain.Section;
-import subway.domain.Station;
+import subway.domain.line.Line;
+import subway.domain.section.Section;
+import subway.domain.station.Station;
 import subway.dto.LineAndStationsResponse;
 import subway.dto.LineRequest;
 import subway.dto.LineResponse;
