+++ src/main/java/subway/application/StationService.java
@@ -2,7 +2,7 @@ package subway.application;
 
 import org.springframework.stereotype.Service;
 import subway.dao.StationEntity;
-import subway.domain.Station;
+import subway.domain.station.Station;
 import subway.dto.StationRequest;
 import subway.dto.StationResponse;
 import subway.repository.StationRepository;
