+++ src/main/java/subway/domain/section/Distance.java
@@ -1,4 +1,4 @@
-package subway.domain;
+package subway.domain.section;
 
 import java.util.Objects;
