+++ src/main/java/subway/domain/station/Station.java
@@ -1,4 +1,4 @@
-package subway.domain;
+package subway.domain.station;
 
 import java.util.Objects;
