+++ src/main/java/subway/dto/LineAndStationsResponse.java
@@ -1,6 +1,6 @@
 package subway.dto;
 
-import subway.domain.Line;
+import subway.domain.line.Line;
 
 import java.util.List;
 import java.util.stream.Collectors;
