+++ src/test/java/subway/application/LineServiceTest.java
@@ -8,8 +8,8 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import subway.dao.LineEntity;
-import subway.domain.Line;
-import subway.domain.Section;
+import subway.domain.line.Line;
+import subway.domain.section.Section;
 import subway.dto.LineAndStationsResponse;
 import subway.dto.LineResponse;
 import subway.dto.StationAddRequest;
