+++ src/test/java/subway/domain/ColorTest.java
@@ -4,6 +4,7 @@ import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
+import subway.domain.line.Color;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
