+++ src/test/java/subway/domain/LinesTest.java
@@ -3,6 +3,8 @@ package subway.domain;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
+import subway.domain.line.Line;
+import subway.domain.section.Section;
 
 import java.util.ArrayList;
 import java.util.List;
