+++ src/test/java/subway/domain/NameTest.java
@@ -4,6 +4,7 @@ import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
+import subway.domain.line.Name;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
