+++ src/test/java/subway/repository/StationRepositoryTest.java
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import subway.domain.Station;
+import subway.domain.station.Station;
 import subway.integration.IntegrationTest;
 
 import static org.assertj.core.api.SoftAssertions.assertSoftly;
